scripts/requirements.txt
pandas == 2.2.3
google-cloud-storage == 2.19.0
google-cloud-aiplatform == 1.79.0
requests == 2.32.3
openpyxl == 3.1.5
matplotlib == 3.10.0
seaborn == 0.13.2
scikit-learn == 1.6.1
kneed == 0.8.5
mlflow == 2.20.1
streamlit == 1.41.1

